// u/src/shared/types/utils/constants.types.ts
export interface PhoneNumberType {
    countryName: string;
    placeholder: string;
    pattern: string;
    countryFlag: string;
}

export type ListDisplayMode = 'list' | 'grid';

export type SocialMediaType =
    | 'instagram'
    | 'tiktok'
    | 'facebook'
    | 'youtube'
    | 'spotify'
    | 'soundcloud'
    | 'press'
    | 'multipromo';